Fix failing test [SCI-3037]
