+++ app/services/model_exporters/model_exporter.rb
@@ -42,6 +42,27 @@ module ModelExporters
       raise NotImplementedError, '#export_to_dir method not implemented.'
     end
 
+    def protocol(protocol)
+      {
+        protocol: protocol,
+        protocol_protocol_keywords: protocol.protocol_protocol_keywords,
+        steps: protocol.steps.map { |s| step(s) }
+      }
+    end
+
+    def step(step)
+      @assets_to_copy.push(step.assets.to_a) if step.assets.present?
+      {
+        step: step,
+        checklists: step.checklists.map { |c| checklist(c) },
+        step_comments: step.step_comments,
+        step_assets: step.step_assets,
+        assets: step.assets,
+        step_tables: step.step_tables,
+        tables: step.tables.map { |t| table(t) }
+      }
+    end
+
     def checklist(checklist)
       {
         checklist: checklist,
